add update balance service with taxes cut

+++ community/src/main/java/com/lawencon/community/dto/balance/BalanceData.java
@@ -1,53 +1,62 @@
 package com.lawencon.community.dto.balance;
 
+import java.math.BigDecimal;
+
 public class BalanceData {
-	
+
 	private String id;
 	private String balanceCode;
-	private Float currentBalance;
+	private BigDecimal currentBalance;
 	private String userId;
 	private Boolean isActive;
 	private Integer version;
-	
+
 	public String getId() {
 		return id;
 	}
+
 	public void setId(String id) {
 		this.id = id;
 	}
+
 	public String getBalanceCode() {
 		return balanceCode;
 	}
+
 	public void setBalanceCode(String balanceCode) {
 		this.balanceCode = balanceCode;
 	}
-	public Float getCurrentBalance() {
+
+	public BigDecimal getCurrentBalance() {
 		return currentBalance;
 	}
-	public void setCurrentBalance(Float currentBalance) {
+
+	public void setCurrentBalance(BigDecimal currentBalance) {
 		this.currentBalance = currentBalance;
 	}
+
 	public String getUserId() {
 		return userId;
 	}
+
 	public void setUserId(String userId) {
 		this.userId = userId;
 	}
+
 	public Boolean getIsActive() {
 		return isActive;
 	}
+
 	public void setIsActive(Boolean isActive) {
 		this.isActive = isActive;
 	}
+
 	public Integer getVersion() {
 		return version;
 	}
+
 	public void setVersion(Integer version) {
 		this.version = version;
 	}
-	
-	
-	
-	
 
 }
